orca/src/requirements.txt
flet==0.28.3
supabase==2.23.0
python-dotenv==1.2.1
